feat: debug log from command line
