+++ packages/cli/src/hooks/init/init-config.ts
@@ -1,6 +1,8 @@
 import type { Hook } from '@oclif/core'
+import log4js from 'log4js'
 
-import { LoginAuthenticator, logManager, defaultLoggingConfig, loadLoggingConfig } from '@smartthings/cli-lib'
+import { LoginAuthenticator } from '@smartthings/cli-lib'
+import { defaultLoggingConfig, loadLoggingConfig } from '../../lib/logger'
 
 
 const hook: Hook<'init'> = async function (opts) {
@@ -9,7 +11,9 @@ const hook: Hook<'init'> = async function (opts) {
 	const defaultLogConfig = defaultLoggingConfig(`${opts.config.cacheDir}/smartthings.log`)
 	const logConfig = loadLoggingConfig(`${opts.config.configDir}/logging.yaml`, defaultLogConfig)
 
-	logManager.init(logConfig)
+	// example of a cli-lib consumer just needing to configure log4js to enable cli logging
+	// no global logManager required
+	log4js.configure(logConfig)
 }
 
 export default hook
