+++ packages/lib/package.json
@@ -28,6 +28,7 @@
 		"directory": "packages/lib"
 	},
 	"dependencies": {
+		"@log4js-node/log4js-api": "^1.0.2",
 		"@oclif/core": "^1.6.4",
 		"@smartthings/core-sdk": "^3.4.1",
 		"@types/eventsource": "^1.1.8",
@@ -40,7 +41,6 @@
 		"inquirer": "^8.2.2",
 		"js-yaml": "^4.1.0",
 		"lodash.at": "^4.6.0",
-		"log4js": "6.3.0",
 		"open": "^8.4.0",
 		"os-locale": "^5.0.0",
 		"qs": "^6.10.3"
@@ -62,6 +62,7 @@
 		"eslint-plugin-jest": "^26.1.3",
 		"jest": "^27.5.1",
 		"jest-extended": "^0.11.5",
+		"log4js": "6.3.0",
 		"mock-stdin": "^1.0.0",
 		"oclif": "^2.6.3",
 		"rimraf": "^3.0.2",
