+++ packages/lib/src/__tests__/api-organization-command.test.ts
@@ -7,7 +7,11 @@ import { ClientIdProvider } from '../login-authenticator'
 jest.mock('os-locale')
 jest.mock('@smartthings/core-sdk')
 jest.mock('../cli-config')
-jest.mock('../logger')
+jest.mock('@log4js-node/log4js-api', () => ({
+	getLogger: jest.fn(() => ({
+		trace: jest.fn(),
+	})),
+}))
 jest.mock('../login-authenticator')
