+++ packages/lib/src/index.ts
@@ -10,7 +10,6 @@ export * from './input'
 export * from './input-builder'
 export * from './io-util'
 export * from './listing-io'
-export * from './logger'
 export * from './login-authenticator'
 export * from './output'
 export * from './output-builder'
